refactor the repository and make testing doubles for testing and pass the first test
