+++ app/src/test/java/com/example/glamora/data/internetStateObserver/FakeInternetStateObserver.kt
@@ -1,0 +1,9 @@
+package com.example.glamora.data.internetStateObserver
+
+import kotlinx.coroutines.flow.Flow
+
+class FakeInternetStateObserver : ConnectivityObserver {
+    override fun observer(): Flow<ConnectivityObserver.InternetState> {
+        TODO("Not yet implemented")
+    }
+}
