+++ app/src/test/java/com/example/glamora/data/repository/RepositoryImplTest.kt
@@ -1,13 +1,24 @@
 package com.example.glamora.data.repository
 
+import androidx.test.core.app.ApplicationProvider
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import app.cash.turbine.test
 import com.apollographql.apollo.ApolloClient
 import com.apollographql.apollo.api.ApolloResponse
 import com.benasher44.uuid.Uuid
 import com.example.ProductQuery
+import com.example.glamora.data.apolloClientHandler.FakeApolloClientHandler
+import com.example.glamora.data.apolloHandler.ApolloClientHandler
+import com.example.glamora.data.contracts.IApolloClient
 import com.example.glamora.data.contracts.RemoteDataSource
+import com.example.glamora.data.contracts.SettingsHandler
+import com.example.glamora.data.firebase.FakeFirebaseHandler
 import com.example.glamora.data.firebase.IFirebaseHandler
 import com.example.glamora.data.internetStateObserver.ConnectivityObserver
+import com.example.glamora.data.internetStateObserver.InternetStateObserver
+import com.example.glamora.data.model.DiscountCodeDTO
+import com.example.glamora.data.network.FakeRemoteDataSource
+import com.example.glamora.data.sharedPref.FakeSharedPrefHandler
 import com.example.glamora.data.sharedPref.SharedPrefHandler
 import com.example.glamora.util.State
 import junit.framework.TestCase.assertEquals
@@ -16,57 +27,61 @@ import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.mockito.Mockito
 import org.mockito.Mockito.mock
+import org.mockito.Mockito.`when`
 import org.mockito.kotlin.whenever
 
-
+@RunWith(AndroidJUnit4::class)
 class RepositoryImplTest {
 
 
-    lateinit var apolloClient : ApolloClient
-    lateinit var remoteDataSource : RemoteDataSource
-    lateinit var sharedPrefHandler : SharedPrefHandler
-    lateinit var connectivityObserver : ConnectivityObserver
-    lateinit var firebaseHandler : IFirebaseHandler
+    private lateinit var apolloClientHandler : IApolloClient
+    private lateinit var remoteDataSource : RemoteDataSource
+    private lateinit var sharedPrefHandler : SettingsHandler
+    private lateinit var connectivityObserver : ConnectivityObserver
+    private lateinit var firebaseHandler : IFirebaseHandler
 
     lateinit var repository: RepositoryImpl
 
     @Before
     fun setUp()
     {
-        apolloClient = mock(ApolloClient::class.java)
-        remoteDataSource = mock(RemoteDataSource::class.java)
-        sharedPrefHandler = mock(SharedPrefHandler::class.java)
-        connectivityObserver = mock(ConnectivityObserver::class.java)
-        firebaseHandler = mock(IFirebaseHandler::class.java)
+        apolloClientHandler = FakeApolloClientHandler()
+        remoteDataSource = FakeRemoteDataSource()
+        sharedPrefHandler = FakeSharedPrefHandler()
+        connectivityObserver = InternetStateObserver(ApplicationProvider.getApplicationContext())
+        firebaseHandler = FakeFirebaseHandler()
 
-        repository = RepositoryImpl(apolloClient, remoteDataSource, sharedPrefHandler, connectivityObserver, firebaseHandler)
+        repository = RepositoryImpl(apolloClientHandler, remoteDataSource, sharedPrefHandler, connectivityObserver, firebaseHandler)
     }
 
     @Test
-    fun `getProducts returns Error when data is null`() = runTest {
-        val mockApolloResponse: ApolloResponse<ProductQuery.Data> = mock()
+    fun `getDiscountCodes returns list of discount codes`() = runTest {
+        val expectedDiscountCodes = listOf(
+            DiscountCodeDTO("1","Summer Sale",5.0),
+            DiscountCodeDTO("2","Summer Sale",10.0),
+            DiscountCodeDTO("3","Summer Sale",15.0),
+            DiscountCodeDTO("4","Summer Sale",20.0),
+            DiscountCodeDTO("5","Summer Sale",25.0),
+        )
 
-        Mockito.`when`(mockApolloResponse.data).thenReturn(null)
-
-
-        Mockito.`when`(apolloClient.query(Mockito.any<ProductQuery>()).execute())
-            .thenReturn(mockApolloResponse)
-
-        Mockito.`when`(mockApolloResponse.data).thenReturn(null)
-
-        repository.getProducts().test {
-
-            assertEquals(State.Loading, awaitItem())
-
-            val errorState = awaitItem() as State.Error
-            assertEquals("No products found", errorState.message)
-
-            cancelAndConsumeRemainingEvents()
+        repository.getDiscountCodes().test {
+            val result = awaitItem() as State.Loading
+            assertEquals(result, State.Loading)
+            val data = awaitItem() as State.Success
+            assertEquals(data.data, expectedDiscountCodes)
+            awaitComplete()
         }
-
     }
+
+
+
+
+
+
+
 
     fun getPriceRules() {
     }
